refactor(logger): abstract away the logger config
