+++ jina/logging/logger.py
@@ -1,78 +1,18 @@
 __copyright__ = "Copyright (c) 2020 Jina AI Limited. All rights reserved."
 __license__ = "Apache-2.0"
 
-import json
 import logging
 import os
 import platform
 import re
 import sys
-from copy import copy
-from logging import Formatter
 from logging.handlers import SysLogHandler
 
 from pkg_resources import resource_filename
 
-from .profile import used_memory
+from . import formatter
 from ..enums import LogVerbosity
-from ..helper import colored, yaml, complete_path
-
-
-class ColorFormatter(Formatter):
-    """Format the log into colored logs based on the log-level. """
-
-    MAPPING = {
-        'DEBUG': dict(color='white', on_color=None),  # white
-        'INFO': dict(color='white', on_color=None),  # cyan
-        'WARNING': dict(color='yellow', on_color='on_grey'),  # yellow
-        'ERROR': dict(color='red', on_color=None),  # 31 for red
-        'CRITICAL': dict(color='white', on_color='on_red'),  # white on red bg
-        'SUCCESS': dict(color='green', on_color=None),  # white on red bg
-    }  #: log-level to color mapping
-
-    def format(self, record):
-        cr = copy(record)
-        seq = self.MAPPING.get(cr.levelname, self.MAPPING['INFO'])  # default white
-        cr.msg = colored(cr.msg, **seq)
-        return super().format(cr)
-
-
-class PlainFormatter(Formatter):
-    """Remove all control chars from the log and format it as plain text """
-
-    def format(self, record):
-        cr = copy(record)
-        if isinstance(cr.msg, str):
-            cr.msg = re.sub(u'\u001b\[.*?[@-~]', '', str(cr.msg))
-        return super().format(cr)
-
-
-class JsonFormatter(Formatter):
-    """Format the log message as a JSON object so that it can be later used/parsed in browser with javascript. """
-
-    KEYS = {'created', 'filename', 'funcName', 'levelname', 'lineno', 'msg',
-            'module', 'name', 'pathname', 'process', 'thread', 'processName',
-            'threadName'}  #: keys to extract from the log
-
-    def format(self, record):
-        cr = copy(record)
-        cr.msg = re.sub(u'\u001b\[.*?[@-~]', '', str(cr.msg))
-        return json.dumps(
-            {k: getattr(cr, k) for k in self.KEYS},
-            sort_keys=True)
-
-
-class ProfileFormatter(Formatter):
-    """Format the log message as JSON object and add the current used memory into it"""
-
-    def format(self, record):
-        cr = copy(record)
-        if isinstance(cr.msg, dict):
-            cr.msg.update({k: getattr(cr, k) for k in ['created', 'module', 'process', 'thread']})
-            cr.msg['memory'] = used_memory(unit=1)
-            return json.dumps(cr.msg, sort_keys=True)
-        else:
-            return ''
+from ..helper import yaml, complete_path
 
 
 class NTLogger:
@@ -87,7 +27,7 @@ class NTLogger:
 
     @staticmethod
     def _planify(msg):
-        return re.sub(u'\u001b\[.*?[@-~]', '', msg)
+        return re.sub(r'\u001b\[.*?[@-~]', '', msg)
 
     def info(self, msg: str, **kwargs):
         """log info-level message"""
@@ -120,7 +60,7 @@ class NTLogger:
             sys.stdout.write(f'W:{self.context}:{self._planify(msg)}')
 
 
-class MySysLogHandler(SysLogHandler):
+class SysLogHandlerWrapper(SysLogHandler):
     """ Override the priority_map :class:`SysLogHandler`
 
     .. warning::
@@ -197,6 +137,7 @@ class JinaLogger:
 
         for h in config['handlers']:
             cfg = config['configs'].get(h, None)
+            fmt = getattr(formatter, cfg.get('formatter', 'PlainFormatter'))
 
             if h not in self.supported or not cfg:
                 raise ValueError(f'can not find configs for {h}, maybe it is not supported')
@@ -204,25 +145,22 @@ class JinaLogger:
             handler = None
             if h == 'StreamHandler':
                 handler = logging.StreamHandler(sys.stdout)
-                if cfg['colored']:
-                    handler.setFormatter(ColorFormatter(cfg['format'].format_map(kwargs)))
-                else:
-                    handler.setFormatter(PlainFormatter(cfg['format'].format_map(kwargs)))
+                handler.setFormatter(fmt(cfg['format'].format_map(kwargs)))
             elif h == 'SysLogHandler':
                 if cfg['host'] and cfg['port']:
-                    handler = MySysLogHandler(address=(cfg['host'], cfg['port']))
+                    handler = SysLogHandlerWrapper(address=(cfg['host'], cfg['port']))
                 else:
                     # a UNIX socket is used
                     if platform.system() == 'Darwin':
-                        handler = MySysLogHandler(address='/var/run/syslog')
+                        handler = SysLogHandlerWrapper(address='/var/run/syslog')
                     else:
-                        handler = MySysLogHandler(address='/dev/log')
+                        handler = SysLogHandlerWrapper(address='/dev/log')
                 if handler:
                     handler.ident = cfg.get('ident', '')
-                    handler.setFormatter(PlainFormatter(cfg['format'].format_map(kwargs)))
+                    handler.setFormatter(fmt(cfg['format'].format_map(kwargs)))
             elif h == 'FileHandler':
                 handler = logging.FileHandler(cfg['output'].format_map(kwargs), delay=True)
-                handler.setFormatter(PlainFormatter(cfg['format'].format_map(kwargs)))
+                handler.setFormatter(fmt(cfg['format'].format_map(kwargs)))
             elif h == 'FluentHandler':
                 try:
                     from fluent import asynchandler as fluentasynchandler
@@ -233,8 +171,8 @@ class JinaLogger:
                                                                port=config['port'], queue_circular=True)
 
                     config['format'].update(kwargs)
-                    formatter = FluentRecordFormatter(config['format'])
-                    handler.setFormatter(formatter)
+                    fmt = FluentRecordFormatter(config['format'])
+                    handler.setFormatter(fmt)
                 except (ModuleNotFoundError, ImportError):
                     pass
