+++ jina/resources/logging.default.yml
@@ -7,15 +7,16 @@ configs:
   FileHandler:
     format: '%(asctime)s:{name:>15}@%(process)2d[%(levelname).1s]:%(message)s'
     output: 'jina-{uptime}.log'
-    type: json  # possible choices ['text', 'json']
+    formatter: JsonFormatter
   StreamHandler:
     format: '{name:>15}@%(process)2d[%(levelname).1s]:%(message)s'
-    colored: true  # show colored logs or not
+    formatter: ColorFormatter
   SysLogHandler:
     ident: # this will be prepend to all messages
     format: '{name:>15}@%(process)2d[%(levelname).1s]:%(message)s'
     host: # when not given then record it locally, /dev/log on linux /var/run/syslog on mac
     port: # when not given then record it locally,  /dev/log on linux /var/run/syslog on mac
+    formatter: PlainFormatter
   FluentHandler:
     # this configuration describes where is the fluentD daemon running and waiting for logs to be emitted.
     # FluentD then will have its own configuration to forward the messages according to its own syntax
