+++ jina/resources/logging.docker.yml
@@ -2,20 +2,20 @@ handlers:
   - StreamHandler
   - SysLogHandler
   - FluentHandler
-level: DEBUG
+level: INFO
 configs:
   FileHandler:
     format: '%(asctime)s:{name:>15}@%(process)2d[%(levelname).1s]:%(message)s'
     output: 'jina-%(uptime)s.log'
-    type: json  # also 'json'
+    formatter: JsonFormatter
   StreamHandler:
     format: '🐳 %(message)s'
-    colored: true
+    formatter: ColorFormatter
   SysLogHandler:
     format: '{name:>15}@%(process)2d[%(levelname).1s]:%(message)s'
     host: # when not given then record it locally
     port: # when not given then record it locally
-    use_socket: false
+    formatter: PlainFormatter
   FluentHandler:
     # this configuration describes where is the fluentD daemon running and waiting for logs to be emitted.
     # FluentD then will have its own configuration to forward the messages according to its own syntax
