+++ jina/resources/logging.profile.yml
@@ -8,6 +8,7 @@ configs:
     format: '{name:>15}@%(process)2d[%(levelname).1s]:%(message)s'
     host: # when not given then record it locally, /dev/log on linux /var/run/syslog on mac
     port: # when not given then record it locally,  /dev/log on linux /var/run/syslog on mac
+    formatter: JsonFormatter
   FluentHandler:
     # this configuration describes where is the fluentD daemon running and waiting for logs to be emitted.
     # FluentD then will have its own configuration to forward the messages according to its own syntax
